<상세 화면> 통신 완료. <홈탭> 리디자인. 서버에서 수정된 부분 반영.

+++ app/src/main/java/com/hyeong/handsomego/get/GetPlaceInfoResponseData.kt
@@ -6,7 +6,7 @@ data class GetPlaceInfoResponseData (
         var place_address : String,
         var place_content : String,
         var place_category : String,
-        var place_star : Int,
+        var place_star : Float,
         var place_pic : String,
         var commentCount : Int
 )
